Completing replacing fixed GridView with gridClass
This commit completes the replacement of `kartik\grid\GridView` with the variable `gridClass` everywhere. rowSelectedClass behavior is slightly different, previously initializing it to `null` or `false` removed it, now only `false` does so. I wanted to preserve the original behavior (for the sake of backwards compatibility) but it seemed the best way, I'd appreciate any suggestions here.